# Speed up unused opens handling for empty results